Enable a DynamoDB Stream
In this commit:
- we updated serverless.yml to  enable a DynamoDB table
- created a lambda function that will be triggered by the db stream from the images table

We are sure the lambda works properly by checking the cloudwatch logs

NOTE: Dynamodb stream is not a streaming service like Knesis or Kafka

We we are to implement full-text search using Knesis, a lambda function called create image wil write write into this Knesis stream. We will have lambda function that will be reading data from this Knesis stream and each data that was added it will create  a new image into the dynamodb table which is nornaml.
In addition, we will have another function that wil read thesame data from the knesis stream and will write the data to elastic search

The MODEL we used for our project, was we just enable UpdateStream from our dynamoDb table, a function writes to the ddb, then a second function reads from the updateStream and replicates them to elastic search