SNS Notifications
Instead of Processing events directly, we subscribed our Lambda function to subscribe to process event from an SNS Topic

- We created an SNS topic
- We connect S3 Bucket to send its events to this SNS topic
- We connected our existing function to this new SNS topic

This allows us to eventually send one S3 event to multiple Lambda functions.
The reason we used Simple Notification Service for our app is because S3 does not allow to send events to multiple targets

SNS is a service that send messages to other services.
Publishers publish messages. A publisher publish message to a topic
Subscribers consume incoming messages. A subscriber receives a message if it is subscribed to a topic. Subscribers can use various protocols like Lambda, HTTP endpoints, SQS. SNS can also send emails, SMS, etc
One topic can have many subscribers. When a message is sent to a topic by one of the publisher, it is send to all subscribers that subscribe to a topic

NOTE: that SNS implements retires as well like the way DDb stream or Knesis does