Resize a File on Upload
Whenever an image is uploaded to the S3 bucket, we will resize it and  put the new resized image into another S3 bucket

This new Lambda function we created:
- loads an uploaded file
- resize file to a smaller image called thumbnail
- save the image to another S3 bucket for thumbnails

One main reason we did not put it back into thesame bucket is because of to avoid a loop where when we upload the resized image another SNS will be send out which will then trigger our lambda function again so we get stuck in that loop. NOT GOOD :( Our new bucket to store thumbnails does not send any trigger event so we are fine

At this point, we have a Bucket send messages to a SNS topic and we have two Lambda functions that are subscribed to that topic. This is a good knowledge