Implement Mock Authentication
We implemented a simple custom authorizer and added that to just two functions createImage and createGroup.

We expected a mock token value in other to authorize the user to run a lambda function.  Ideally it will be a real JWT token from Auth0 service

We used Auth0 just to show that we can use a non AWS service in our application.
AWS has an auth service called AWS Cognito but we did not use it

More CustomAuthorizer in go
https://github.com/serverless/examples/blob/master/aws-golang-auth-examples/functions/auth2/main.go
https://gist.github.com/praveen001/1b045d1c31cd9c72e4e6638e9f883f83
https://github.com/aws/aws-lambda-go/blob/master/events/README_ApiGatewayCustomAuthorizer.md