Configure Canary Deployment
In summary the Process of Canary deployment is migrate traffic to the new version gradually and rollback if any problems occur

- In this commit, we had a previous version of our lambda from the last commit was was saved in aws lambda by default(eg v50).
- Now we made a small change in our code and deployed this change. This lead to a new version to be created v51
- Now we wanted to share traffic between these two versions of our Lambda, and we do this with the help of an alias. We created an alias called Live(You can name it whatever like Prod). This alias point at thesame time to those two versions and we assign 10% percent probability of routing traffic to the new version v51 every 1 minute and 90% to v50
- NOTE that you can configure setting to rollback and route all traffic to the old stable version if an error code was returned in the canary(new) version and if thee is no alarm yet, we will keep redirecting more and more percentage of traffic of our requests to the canary until all requests are now directed to the canary version  https://www.serverless.com/plugins/serverless-plugin-canary-deployments

APIGateway Points to an Alias, so when a request comes in through APIGateway, it goes to either on of the Lambda function

Alias is like a reference that points to a specific lambda function. So instead of calling directly a lambda function with a specific version, a client can call a lambda function by its alias.
AWS Lambda also supports an alias that points to two versions of thesame function and when this alias is called, the invocation goes to either one of those two versions. The version that will be called is selected at random but we can assign a probability of an invocation of each version.

NOTE: When do run sls deploy -v, the canary deployment takes a while depending on what you set on it, so you can just start testing the function as the "UPDATE_IN_PROGRESS - AWS::CodeDeploy::DeploymentGroup" is running!