Optimize Sizes of our functions
Before this commit, when we look at all our functions in the AWS console, al the functions had thesame size (47.5MB).
This is because, serverless packages all the dependencies that are used by atleast one function and create a single zip package and thesame zip package is used for all functions the only difference between our functions is to what handler to process incoming events

After the changes in this commit at the sls config file, we now have smaller functions that should give us lower cold start time but we have to pay it for longer compilation time. So, judge the trade off yourself!