Implement Pagination
The GET/groups API have two parameters
limit: which is the maximum number of items an API should return
nextKey: which is an information that is returned by our API call so that we can use it to get more data on the next request

Query or Scan operations in DynamoDB can only return 1MB of data. It is a hard limit so it is important to use pagination if you want to read more than 1MB of data out of a dynamoDb table.

DynamoDb returns "LastEvaluatedKey" parameter and if it is not equal to null, it means that there is still more data to return. So, on the next query, we can pass this parameter and dynamoDb will return the next batch of items

A good read on when you are querying or scanning items with pagination on an index table
https://www.talentica.com/blogs/dynamo-db-pagination/